Added test and changes to write_nwb method except blocks

+++ allensdk/brain_observatory/behavior/write_nwb/__main__.py
@@ -53,7 +53,8 @@ def write_behavior_ophys_nwb(session_data: dict,
         os.rename(nwb_filepath_inprogress, nwb_filepath)
         return {'output_path': nwb_filepath}
     except Exception as e:
-        os.rename(nwb_filepath_inprogress, nwb_filepath_error)
+        if os.path.isfile(nwb_filepath_inprogress):
+            os.rename(nwb_filepath_inprogress, nwb_filepath_error)
         raise e
